# Use a LinkedBlockingQueue in PacketQueue.