This allows us to pass a capacity of Integer.MAX_VALUE for unbounded queues.