@JonathanLennox 2021-08-03
The PacketQueueBenchmarkTests don't seem to show appreciable performance differences with this change.  Testing on my perf machine:

On master:
```
CachedThreadPerQueuePool : avg = 676 ms, std_dev = 83 ms
CachedThreadPerQueuePoolWithStatistics : avg = 722 ms, std_dev = 92 ms
FixedSizeCPUBoundPool : avg = 630 ms, std_dev = 62 ms
FixedSizeCPUBoundPoolWithStatistics : avg = 668 ms, std_dev = 59 ms
ForkJoinCPUBoundPool : avg = 627 ms, std_dev = 56 ms
ForkJoinCPUBoundPoolWithStatistics : avg = 670 ms, std_dev = 58 ms
ThreadPerQueuePool : avg = 794 ms, std_dev = 88 ms
ThreadPerQueuePoolWithStatistics : avg = 844 ms, std_dev = 81 ms
```

With this PR:
```
CachedThreadPerQueuePool : avg = 669 ms, std_dev = 64 ms
CachedThreadPerQueuePoolWithStatistics : avg = 724 ms, std_dev = 61 ms
FixedSizeCPUBoundPool : avg = 656 ms, std_dev = 65 ms
FixedSizeCPUBoundPoolWithStatistics : avg = 693 ms, std_dev = 66 ms
ForkJoinCPUBoundPool : avg = 650 ms, std_dev = 56 ms
ForkJoinCPUBoundPoolWithStatistics : avg = 695 ms, std_dev = 59 ms
ThreadPerQueuePool : avg = 828 ms, std_dev = 88 ms
ThreadPerQueuePoolWithStatistics : avg = 876 ms, std_dev = 83 ms
```

Alternately, we could use a LinkedBlockingQueue only when capacity is Integer.MAX_VALUE.